app/pod-observer/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
Requests==2.32.3
slackclient==2.9.4
gunicorn==23.0.0